json/tools: WIP cell phone to drain charges while on.
...and drain them very fast while the "flashlight app" is on.

Should eventually close #16634.

Two `transform`s for one item doesn't seem to be possible (game
only allows first action), so turning on/off uses `delayed_transform`.

That "simulates" the phone booting, but doesn't seem to work as
intended - the item is transformed instantly. Is the `transform_age` in
turns or seconds?.. `doc/JSON_INFO.md` suggests it's in turns.