vehicle: More iterations for collision loop
For slow-moving vehicles, especially heavy ones, impulse delta is reasonably negligible,
but the velocity is also effectively multiplied by 0.9.

The reason for this is unknown to me. Perhaps to guarantee an exit?.. Checking git blame
and digging for when the code was added might yield an answer, but the whole section
uses a scalar calculation (i.e. it assumes only one vehicle is moving pre-collision, and
that the velocities can be scalars instead of vectors).

As a quick fix, just increase the multiplier.